om/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
djangorestframework==3.6.3
docutils==0.14
gunicorn==19.7.1
idna==2.6
Markdown==2.6.9
psycopg2==2.7.3
Pycco==0.5.1
Pygments==2.2.0
pystache==0.5.4
pytz==2017.2
requests==2.18.4
smartypants==2.0.0
urllib3==1.22
